Moved table document with interception to packages/components/table-document.
