+++ packages/components/table-document/package.json
@@ -55,6 +55,7 @@
     "@microsoft/eslint-config-fluid": "^0.14.0",
     "@microsoft/fluid-build-common": "^0.14.0",
     "@microsoft/fluid-local-test-utils": "^0.14.0",
+    "@microsoft/fluid-local-web-host": "^0.14.0",
     "@types/debug": "^0.0.31",
     "@types/mocha": "^5.2.5",
     "@types/node": "^10.14.6",
