+++ packages/components/table-document/tsconfig.json
@@ -8,7 +8,8 @@
         "strict": true,
         "strictNullChecks": false,
         "rootDir": "./src",
-        "outDir": "./dist"
+        "outDir": "./dist",
+        "types": [ "mocha" ]
     },
     "include": [
         "src/**/*"
