+++ packages/framework/dds-interceptions/package.json
@@ -49,7 +49,6 @@
     "temp-directory": "nyc/.nyc_output"
   },
   "dependencies": {
-    "@fluid-example/table-document": "^0.14.0",
     "@microsoft/fluid-map": "^0.14.0",
     "@microsoft/fluid-merge-tree": "^0.14.0",
     "@microsoft/fluid-runtime-definitions": "^0.14.0",
@@ -58,8 +57,6 @@
   "devDependencies": {
     "@microsoft/api-extractor": "^7.7.2",
     "@microsoft/eslint-config-fluid": "^0.14.0",
-    "@microsoft/fluid-aqueduct": "^0.14.0",
-    "@microsoft/fluid-local-web-host": "^0.14.0",
     "@microsoft/fluid-test-runtime-utils": "^0.14.0",
     "@types/debug": "^0.0.31",
     "@types/diff": "^3.5.1",
